src/main/java/com/example/challengetechnique/services/impl/UserService.java
package com.example.challengetechnique.services.impl;

import com.example.challengetechnique.models.User;
import com.example.challengetechnique.repositories.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class UserService {

    private final UserRepository userRepository;

    public User profile(String username) {
        return userRepository.findUserByUsername(username).orElseThrow(() -> new UsernameNotFoundException("No user found"));
    }

}
